requirements.txt
Django>=3.0,<3.2
uwsgi>=2.0,<2.1
dj-database-url>=0.5,<0.6
psycopg2>=2.8,<2.9
django-oauth-toolkit>=1.5,<1.6
djangorestframework>=3.12,<3.13
RandomWords>=0.2,<0.3
Pillow>=7,<8
django-cors-headers>=3.7,<3.8